refactor: standardize models and rename types folder
