+++ scorecard-webapp/src/models/CourseSummary.ts
