+++ scorecard-webapp/src/models/Hole.ts
@@ -1,4 +1,4 @@
-export type Hole = {
+export interface Hole {
   number: number;
   par: number;
-};
+}
